reorganize files and schema
