+++ server/api/utils/session.ts
