+++ server/database/models/clientModel.ts
@@ -2,27 +2,38 @@ import { Document, model, Schema, Types } from 'mongoose'
 
 interface ClientInterface {
   _id: Types.ObjectId
-  clientId: string
+
+  // Properties
   fullName: string
+
+  // Status
   deleted: boolean
+
+  // Timestamps
+  createdAt: Date
+  updatedAt: Date
 }
 
 type ClientDocument = ClientInterface & Document;
 
 const ClientSchema = new Schema({
-  clientId: {
-    type: String,
-  },
+  // Properties
   fullName: {
     type: String,
     required: true,
     trim: true
   },
+
+  // Status
   deleted: {
     type: Boolean,
     required: true,
     default: false
   }
+
+}, // Options
+{
+  timestamps: true
 })
 
 const Client = model('Client', ClientSchema)
