+++ server/database/models/newSchema.txt
@@ -1,0 +1,56 @@
+interface RequestGroupInterface {
+    _id: Types.ObjectId,
+    name: string,
+    description: string,
+    dateUpdated: Date,
+    deleted: boolean,
+    image: string,
+    requestTypes: [
+        {
+            _id: Types.ObjectId, ref: 'RequestType',
+            //deleted: boolean,
+            //dateUpdated: Date,
+            //name: string,
+            //numOpen: number
+        }
+    ]
+  }
+
+  interface RequestTypeInterface {
+    _id: Types.ObjectId,
+    requestGroup: Types.ObjectId,
+    dateUpdated: Date,
+    deleted: boolean,
+    name: string,
+    numOpen: number,
+    requests: [
+      {
+          _id: Types.ObjectId, ref: 'Request',
+          //deleted: boolean,
+          //fulfilled: boolean,
+          //dateUpdated: Date,
+          //dateCreated: Date,
+          //dateFulfilled: Date,
+          //quantity: number 
+      }
+  ]
+  }
+
+  interface RequestInterface {
+    _id: Types.ObjectId
+    requestType: Types.ObjectId
+    client: Types.ObjectId
+    dateUpdated: Date
+    dateCreated: Date
+    dateFulfilled: Date
+    deleted: boolean
+    fulfilled: boolean
+    quantity: number 
+  }
+
+  interface ClientInterface {
+    _id: Types.ObjectId
+    clientId: string
+    fullName: string
+    deleted: boolean
+  }
